Fix diff-hl not showing up and removing symbols.
Removed symbols like +,-,= etc by using customize, and turned off
diff-hl-margin-mode because it would turn off the display completely,
and without my use-package setup for diff-hl it worked, so I bisected
and figured out that was the call doing it, so I left the margin mode
off. Seems to be working as intended now. Not sure when to turn on
margin mode though...